Hair Dyeing App
This app is a combination of a timer for bleaching your hair and a
timer for coloring your hair. It also has a shopping list that you can
add items to so that you can remember to replace your inventory after
you just used it. If your like me then this might come in handy when
dyeing your hair into Viewtiful colors. Viewtiful with a capital V.